refactor: extract status overlay and settings form
